automation/chatbot/requirements.txt
jupyterlab
python-telegram-bot